requirements.txt
tiktoken
huggingface_hub
langchain
streamlit
google-generativeai
langchain-google-genai
python-dotenv